Updated Vitest to 1.5.0, refactored some tests.
