+++ frontend/src/__tests__/components/TimerAutomationCard.test.jsx
@@ -3,7 +3,8 @@ import { BrowserRouter as Router } from "react-router-dom";
 import { test, expect } from "vitest";
 import { TimerAutomationCard } from "../../components/Automations/TimerAutomationCard.jsx";
 import { I18nextProvider } from "react-i18next";
-import i18n from "../../i18n.js";
+import { AuthProvider } from "../../context/AuthContext.jsx";
+import i18n from "../../i18n-test.js";
 
 test("TimerAutomationCard renders correctly", () => {
   const mockAutomation = {
@@ -35,11 +36,13 @@ test("TimerAutomationCard renders correctly", () => {
     ],
   };
   render(
-    <Router>
-      <I18nextProvider i18n={i18n}>
-        <TimerAutomationCard automation={mockAutomation} />
-      </I18nextProvider>
-    </Router>
+    <AuthProvider>
+      <Router>
+        <I18nextProvider i18n={i18n}>
+          <TimerAutomationCard automation={mockAutomation} />
+        </I18nextProvider>
+      </Router>
+    </AuthProvider>
   );
 
   const cardElement = screen.getByTestId("automation-card");
